LPS-115438 Extract fix language value logic into portal kernel and deprecate old logic, so that it can be used by ResourceBundleUtil
