+++ portal-kernel/src/com/liferay/portal/kernel/language/LanguageBuilderUtil.java
@@ -12,18 +12,30 @@
  * details.
  */
 
-package com.liferay.portal.tools;
+package com.liferay.portal.kernel.language;
 
 /**
- * @author Brian Wing Shun Chan
- * @deprecated As of Athanasius (7.3.x), with no direct replacement
+ * @author Tina Tian
  */
-@Deprecated
-public class LangBuilder {
+public class LanguageBuilderUtil {
 
 	public static final String AUTOMATIC_COPY = " (Automatic Copy)";
 
 	public static final String AUTOMATIC_TRANSLATION =
 		" (Automatic Translation)";
+
+	public static String fixValue(String value) {
+		if (value.endsWith(AUTOMATIC_COPY)) {
+			value = value.substring(
+				0, value.length() - AUTOMATIC_COPY.length());
+		}
+
+		if (value.endsWith(AUTOMATIC_TRANSLATION)) {
+			value = value.substring(
+				0, value.length() - AUTOMATIC_TRANSLATION.length());
+		}
+
+		return value;
+	}
 
 }
