+++ portal-kernel/src/com/liferay/portal/kernel/language/packageinfo
@@ -1,1 +1,1 @@
-version 8.1.0
+version 8.2.0
